Update data gen to support child tables.
In some of the tpcds tables, there is a parent/child relationship. This means that
the child table cannot be generated independently and is output (optionally) as part
of generating the parent table. This benchmark requires the output being written to
stdout. When the tpcds generator runs with stdout as the output, it cannot support
child tables.

The dsdgen tool has been updated to support this and can optionally output the data
with the table prefix. For example, in this mode the output is:

store_sales|<row>
store_sales|<row>
store_returns|<row>
store_sales|<row>

The rows are interleaved but can be filtered for. This patch adds the filtering.